+++ packages/log/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @idearium/log
 
+## Unreleased
+
+-   Refined the automatic `sourceLocation` result.
+
 ## v1.0.0-beta.5 - 2021-05-05
 
 -   Automated the inclusion of the `sourceLocation` property.